task/views/Image.py
from rest_framework import viewsets
from task.models import Image
from task.serializers import ImageSerializer


class ImageViewSet(viewsets.ModelViewSet):
    queryset = Image.objects.all()
    serializer_class = ImageSerializer
