pyproject.toml
[project]
name = "whys-task-be"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Thyrst", email = "thyrst@seznam.cz" }
]
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/whys_task_be"]
